Fixed problems with pull request #44 (Issues #8 #9)
